pyproject.toml
[tool.poetry]
name = "sbti-finance-tool-api"
version = "2.0.0"
description = "REST API for the SBTi Finance Temperature Alignment tool — temperature scoring, portfolio coverage, and what-if scenario analysis."
authors = ["sbti <finance@sciencebasedtargets.org>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
]
keywords = ['Climate', 'SBTi', 'Finance', 'API']

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/ScienceBasedTargets/SBTi-finance-tool-api/issues"
"Source Code" = "https://github.com/ScienceBasedTargets/SBTi-finance-tool-api"

[tool.poetry.dependencies]
python = ">=3.11,<4"
fastapi = ">=0.115,<1"
uvicorn = {extras = ["standard"], version = ">=0.32,<1"}
python-multipart = ">=0.0.12"
sbti-finance-tool = ">=1.2.5"
pandas = ">=1.5.3,<4.0.0"
openpyxl = ">=3.1"
numpy = ">=1.21,<3"

[tool.poetry.group.dev.dependencies]
pytest = ">=8"
httpx = ">=0.27"
ruff = ">=0.8"

[tool.pyright]
venv = ".venv"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
